refactor(dsh-mcp-manager): 阶段5目录域成形——src/catalog/ 六文件+interface.ts、catalogViewFor 迁出、routes 类型环修复（Part of #664）
Summary:
- 新建 src/catalog/（D10 门面 interface.ts）：entries.ts（条目/摘要/渲染/转义/
  定位）、digest.ts（digestCatalogEntries）、history.ts（catalogHistory）、
  injection.ts（resolveCatalogInjection）、search.ts（scoreTool/searchCatalog/
  searchCatalogMulti/listCatalog/findToolDetail/isCatalogFresh/boundCatalogTools
  自 middleware-utils.ts 迁出，B10 截断口径随迁不动）、cache-view.ts
  （catalogViewFor 自 manager.ts 迁出，宿主最小面 CatalogViewHost + 工厂闭包
  持有 diskCatalogSummaryCache；catalogCacheFile 随迁）
- catalog.ts 删除（引用全切后 git rm）；middleware-utils.ts 仅留策略裁决与
  禁用表域；manager.ts catalogViewFor 变薄桥接（构造时 makeCatalogViewFor 闭合
  host 读取器）
- routes.ts:21 类型环修复：ClientUiConfig 改从 types.ts 取（消 index.ts 环）
- 静态面同步：mutation-topology runtime 段移除 catalog.ts（C4 变异盲区明示），
  gen-stryker-conf 重生成

Design decisions:
- 迁移模式=新文件承载实现+旧位置删定义；diskCatalogSummaryCache 用工厂闭包
  持有（模块级 Map 会多实例串状态）
- host 最小面用读取器（getMiddleware/getMiddlewareMode）——热切换后中间层
  实例与模式变更需取最新引用，闭包捕获引用会失效
- 阶段 5 核对 B10 已在阶段 2 修复态（truncated 用 > 非 >=），本阶段只随迁

Related issue #664
