+++ packages/dsh-mcp-manager/src/apply-config.ts
@@ -8,7 +8,7 @@
 
 import type { Context } from "@deepseek-ai/cordis";
 import { installSettingsNamespace } from "../../../shared/settings-namespace.js";
-import { DEFAULT_ANNOUNCE_CATALOG, DEFAULT_CATALOG_MAX_ENTRIES } from "./catalog.ts";
+import { DEFAULT_ANNOUNCE_CATALOG, DEFAULT_CATALOG_MAX_ENTRIES } from "./catalog/interface.ts";
 import { Config, DEFAULT_ENHANCE_EMPTY_DESCRIPTIONS } from "./config-schema.ts";
 import { DEFAULT_RESULT_TRUNCATE_BYTES } from "./supervisor.ts";
 import { normalizeMiddlewareMode } from "./workspace/interface.ts";
