+++ packages/dsh-mcp-manager/src/manager.ts
@@ -9,14 +9,14 @@
  */
 
 import { mkdir, readFile, rename, writeFile } from "node:fs/promises";
-import { existsSync, statSync } from "node:fs";
+import { existsSync } from "node:fs";
 import { dirname, join } from "node:path";
 import type { ServerResponse } from "node:http";
 import type { SseHub } from "../../../shared/sse-hub.js";
 import type { Context, LoggerService } from "@deepseek-ai/cordis";
 import type { ServerConfig, ClientUiConfig } from "./types.ts";
 import type { ToolDefinition } from "@deepseek-ai/dsh-tools";
-import type { CatalogCache } from "./catalog.ts";
+import type { CatalogCache } from "./catalog/interface.ts";
 import { normalizeServer } from "./normalize.ts";
 import { normalizeUiConfig, buildConfigUiPatch } from "./config-schema.ts";
 import { McpStore } from "./store.ts";
@@ -29,7 +29,8 @@ import {
   findProjectRoot,
   normalizedProjectRoot,
 } from "./workspace/interface.ts";
-import { catalogCacheFile, summarizeToolDescriptions } from "./catalog.ts";
+import { catalogCacheFile, summarizeToolDescriptions, makeCatalogViewFor } from "./catalog/interface.ts";
+import type { CatalogViewResolver } from "./catalog/interface.ts";
 import {
   McpMiddleware,
   msgOf,
@@ -37,7 +38,6 @@ import {
   loadUserState,
   saveUserState,
   catalogCacheFileFor,
-  readCatalogServerFromDisk,
   loadDisabledTools,
   saveDisabledTools,
 } from "./middleware.ts";
@@ -87,6 +87,8 @@ export class McpManager {
   runtimeRegistry: Map<string, ServerConfig>;
   /** registerServer 串行队列（防 reconcileBusy 吞注册；多插件并发注册排队）。 */
   private registerQueue: Promise<void>;
+  /** 注入端目录缓存视图解析器（catalog/cache-view.ts 工厂闭包；含 mtime 缓存）。 */
+  private catalogViewResolver: CatalogViewResolver;
 
   constructor(ctx: Context, store: McpStore) {
     this.ctx = ctx;
@@ -106,6 +108,14 @@ export class McpManager {
     // 目录缓存：serverName → { summary }（磁盘持久化，digest 的稳定数据源）。
     this.catalogCache = new Map();
     this.catalogCachePath = catalogCacheFile();
+    // 目录视图解析器：host 面用读取器（middleware/模式热切换后取最新引用）。
+    const self = this;
+    this.catalogViewResolver = makeCatalogViewFor({
+      getCatalogCache: () => self.catalogCache,
+      getMiddleware: () => self.middleware,
+      getMiddlewareMode: () => self.middlewareMode,
+      catalogCachePathFor: (root) => self.catalogCachePathFor(root),
+    });
     this.uiConfigSource = () => ({ position: "top-right", offsetX: 8, offsetY: 8, blankY: 40 });
     this.middlewareMode = "off";
     this.middleware = undefined;
@@ -377,29 +387,10 @@ export class McpManager {
     return servers;
   }
 
-  /** 磁盘 last-good 摘要缓存（root → server → { mtimeMs, summary }）；
-   * #569 防 pre-step 每轮重复读盘解析（目录文件最大 ~256KB，JSON parse 有成本）。 */
-  private diskCatalogSummaryCache = new Map<string, Map<string, { mtimeMs: number; summary: string | undefined }>>();
-
   /**
-   * 合成注入端目录缓存视图（#569 修复核心）：以 B（supervisor 摘要缓存）为基底，
-   * 逐服务器按「scope + 模式」用中间层目录摘要覆盖——修复「middleware 采集的
-   * 工具摘要注入端读不到」：
-   * - 用户手写 server.description 由 composeCatalogEntries 处理（优先级最高），
-   *   不在此视图内；
-   * - 本视图只负责注入端的 ②（中间层目录摘要）与 ③（B 缓存摘要）两级回退；
-   * - **判 middlewareMode 而非 middleware 实例**（评审修正①）：热切换 off 后
-   *   middleware 实例与 units 残留（apply.setMiddlewareMode 只卸载 ws_mcp_* 工具），
-   *   判实例会错误读中间层目录。
-   *
-   * root 映射（与 middlewareTakes 同口径 + 热切换残留兜底，评审修正②③）：
-   * - mode=off → 全部走 B（无中间层目录语义）；
-   * - scope=project → 该 cwd 归一化项目 root 的单元 → 无则磁盘 last-good；
-   * - scope=global → @global 单元/磁盘（all 模式权威；project 模式热切换残留/
-   *   上次 all 会话的 last-good 兜底）→ 仍无则保留 B。
-   * 严格按 scope 解析、不跨 scope 混配同名服务器（防不同配置被错配）。
-   * 单元不存在时读磁盘 last-good（带 mtime 缓存），**不主动 projectUnitFor**
-   * ——pre-step 无连接副作用。
+   * 合成注入端目录缓存视图（#569 修复核心）：实现在 catalog/cache-view.ts
+   * （makeCatalogViewFor 工厂闭包，含磁盘 last-good mtime 缓存），本方法为
+   * 薄桥接（C-DIR：catalogViewFor 迁出 catalog 域，宿主最小面）。
    * @param cwd 会话 cwd（与 catalogServersFor 同源解析项目 root）。
    * @param servers catalogServersFor 的输出（name → { server, scope }）。
    * @returns CatalogCache 形态视图（name → { summary }），可直接喂 composeCatalogEntries。
@@ -408,60 +399,7 @@ export class McpManager {
     cwd: string | undefined,
     servers: Map<string, { server: ServerConfig; scope: string }>,
   ): Promise<CatalogCache> {
-    const view: CatalogCache = new Map();
-    for (const [name, entry] of this.catalogCache) view.set(name, { summary: entry.summary });
-    const mw = this.middleware;
-    if (mw === undefined || this.middlewareMode === "off") return view;
-    // 项目 root 只解析一次（所有 project scope 服务器共用；空 cwd → 无项目单元）。
-    const cwdRoot = cwd === undefined || cwd === null || cwd === "" ? undefined : await normalizedProjectRoot(cwd);
-    for (const [name, { scope }] of servers) {
-      if (name === "") continue;
-      // root 解析：project scope → 项目 root；global scope → @global（all 权威，
-      // project 模式残留兜底）。跨 scope 不混配（同名项目级/全局是两份配置）。
-      const root = scope === SCOPE_PROJECT ? cwdRoot : MIDDLEWARE_GLOBAL_ROOT;
-      if (root === undefined) continue;
-      const summary = await this.middlewareCatalogSummary(mw, root, name);
-      if (summary !== undefined) view.set(name, { summary });
-    }
-    return view;
-  }
-
-  /** 查中间层单服务器目录摘要：内存单元优先，单元缺失/无该服务器 → 磁盘
-   * last-good 兜底（带 mtime 缓存，防 pre-step 每轮读盘）。返回 undefined
-   * 表示中间层无此服务器数据（调用方保留 B 兜底）。 */
-  private async middlewareCatalogSummary(mw: McpMiddleware, root: string, name: string): Promise<string | undefined> {
-    const unit = mw.units.get(root);
-    const catalog = unit?.catalog.get(name);
-    const tools = catalog?.tools;
-    if (tools !== undefined && tools.size > 0 && catalog?.unavailable === undefined) {
-      // 内存单元（已含磁盘 last-good：单元创建时 loadCatalogCache 载入）→ 直接聚合。
-      return summarizeToolDescriptions(tools as Map<string, { description?: unknown }>);
-    }
-    // 内存无该服务器数据（单元未建 / 条目缺失 / 发现失败）→ 磁盘 last-good 兜底。
-    const file = this.catalogCachePathFor(root);
-    const cachedRoot = this.diskCatalogSummaryCache.get(root);
-    const cached = cachedRoot?.get(name);
-    let mtimeMs = 0;
-    try {
-      if (!existsSync(file)) return undefined;
-      const stat = statSync(file);
-      mtimeMs = stat.mtimeMs;
-    } catch {
-      return undefined; // 文件消失/不可读 → 无磁盘兜底
-    }
-    if (cached?.mtimeMs === mtimeMs) return cached.summary;
-    const persisted = await readCatalogServerFromDisk(file, name);
-    if (persisted === undefined) return undefined;
-    const summary = summarizeToolDescriptions(
-      new Map(persisted.tools.map((tool) => [tool.name, { description: tool.description }])),
-    );
-    let rootCache = this.diskCatalogSummaryCache.get(root);
-    if (rootCache === undefined) {
-      rootCache = new Map();
-      this.diskCatalogSummaryCache.set(root, rootCache);
-    }
-    rootCache.set(name, { mtimeMs, summary });
-    return summary;
+    return this.catalogViewResolver(cwd, servers);
   }
 
   /**
